Changed back due to Bukkit limit.

+++ src/main/java/ch/njol/skript/effects/EffPotion.java
@@ -108,7 +108,7 @@ public class EffPotion extends Effect {
 			final Timespan dur = duration.getSingle(e);
 			if (dur == null)
 				return;
-			d = (int) (dur.getTicks_i() >= Long.MAX_VALUE ? Long.MAX_VALUE: dur.getTicks_i());
+			d = (int) (dur.getTicks_i() >= Integer.MAX_VALUE ? Integer.MAX_VALUE: dur.getTicks_i());
 		}
 		for (final LivingEntity en : entities.getArray(e)) {
 			for (final PotionEffectType t : ts) {
